Beginner JS project
This is An HTML CSS and JavaScript  project that wound expose beginners to the manipulation of CSS properties  with JavaScript.
This project runs a function that alters the CSS top property to move between  multiple slides either upwards or downwards when the slider button is clicked. The transition is also made with  CSS transition .